Use puma instead of thin
Just to troll Oleg... *huk huk huk*

But seriously, puma has built in functionality for both phased restarts
and hot restarts, which makes it very easy to deal with zero downtime
deploys by sending a USR1 or USR2 to the process.  This will be helpful
in future commits, but making a transition now so it is the only
webserver we are using in the project.

And again, to troll Oleg...