Add intial webhook_receiver and webhook processor
This commit adds two things:

* A WebhookProcessor worker, who's sole job is to route webhooks from
  github to their respective jobs to be processed.  With this setup,
  this means that a webhook will be processed by sidekiq twice:  Once to
  receive the webhook and put it in the queue, from there, it will be
  processed and determined which main job should be created after that.
  The jobs for the WebhookProcessor should be short and quick, allowing
  it to process a high volume (not that it is terribly critical for
  speed once it is at this point). In it's current state, this is
  basically outputing a message to the logs saying it received the
  message (correct implementation will be done in the future)
* A WebhookReceiver rack app, which is a simplistic basic auth'd
  api that responds to the `/payload` route, and creates a
  WebhookProcessor job from the data received.

The code for the WebhookProcessor is still a stub, but it is used by the
WebhookReceiver to queue a job, so it has been implemented bare bones
for now to allow the WebhookReceiver to function as it will in the
future.  The rest of this commit message will explain the specific
details of how the WebhookReceiver works, is run, and can be deployed
and developed on.

Running
-------
The process can be triggered via foreman (it has been added to the
Procfile.tmpl as webhook_proxy) like the rest of the application, and if
you just want a minimal setup for testing and working with the webhook
code here, you can run with the following Procfile:

    webhook_proxy:   bundle exec puma webhook_receiver.ru
    sidekiq:         bundle exec sidekiq -q miq_bot
    redis: redis-server /usr/local/etc/redis.conf

And the following command:

    $ foreman start -m redis=1,sidekiq=1,webhook_proxy=1

Redis being optional if you are already running that in the background.

In a separate terminal, you can actually restart the running process
without a shutdown using `pumactl`:

    $ pumactl -P tmp/webhook_processor.pid phased-restart

Any code changes made to the webhook_processor.ru, or any loaded libs
(even from bundler) will be reloaded, but the existing process will
remain unchanged.  In the current setup, we are not running in clustered
mode, so this a "hot restart", and while the socket is remained open
during the process of the restart, and requests made during this time
will hang.  We can change the configuration later, but the process in
it's current state only takes 25ish MB, and is less than a second to
restart.

Routes
------
The WebhookReceiver's last couple of lines map the two routes that are
available to it:

* The '/' (Heartbeat) can be used to just confirm that app is running by
  any external monitoring tool.  This can be made more advanced in the
  future, but this was simply to provide a route for status checks.
  This Heartbeat class also will match any other route given to it
  (minus '/payload'), so it 404s if it receives anything else.
* The 'payload' route that receives webhooks from github, confirms the
  bare minimum that they are valid, creates a Sidekiq job for them to be
  sorted further, and returns a 200.  If anything errors, a 500 is
  returned.

From the docs, you can see and example payload:

https://developer.github.com/webhooks/#payloads

And the validator currently confirms the existence of the two headers
("X-GitHub-Event" and "X-GitHub-Delivery"), and that the body of the
request is not empty.  With that, it will create a Sidekiq Job, and
return a 200.

Auth
----
Currently, basic auth is implemented with hard coded username and
password of `admin` and `smartvm` respectively.  It can be switched to
use `Settings` in the future, or use X-Hub-Signature if we would prefer
to used Github's built in verification mechanisms.

Making a fake Webhook Request (testing it out)
----------------------------------------------
You can run the following curl command:

  curl -H "X-Github-Delivery: 72d3162e-cc78-11e3-81ab-4c9367dc0958" \
       -H "X-GitHub-Event: issues" \
       -H "Content-Type: application/json" \
       -X POST \
       -d '{ "action": "fake" }' \
       "admin:smartvm@localhost:9292/payload"

Do it Live!
-----------
Or, if you are able to expose your running webhook process to the web,
you can setup a sample hook on your own personal `miq_sandbox` repo, and
make sure the following are set:

* The url is something like:
    http://admin:smartvm@XXX.XXX.XXX.XXX:9292/payload
* Content type is "application/json"
* Send Everything
* and Active

You should get a sample hook right away (and you can resend it if it
fails), and other actions done to the repo will also receive events.