[FIXUP] only run validations on @test_repos
Since we might want to use gems changes from outside the org, I think
this makes sense to allow anything for `@repos`.

For `@test_repos`, however, the most convincing case for this has been:

  > Hmmm... I'm not sure the bot should be the one to make that
  > determination. That is, if I create a repo in my personal fork, say a
  > new provider plugin I eventually want to introduce into manageiq, and
  > it does have rake setup appropriately, then I feel the bot should be
  > allowed to run it.

But I feel like this is a rare case, and I think having the bot avoid a
botched run is more valuable 99 times out of 100 compared to the few
cases where a new provider plugin would make use of this case.

Also, this would only be an issue with having the test suite of that new
provider plugin be run by the bot, which can just be as easily don't
within the new provider repo as a separate PR, if integrating some extra
PR changes was necessary to get a passing suite.  Again, since this is a
corner case, I would rather have this feature in the bot then try
cripple the bot for a rare situation.