[GithubService] Add username lookup
Adds a lookup for usernames to check to see if they are valid.

- If a valid username is provided, it returns the ID of the user.
- If username is invalid, it sets the key to `nil`

Only valid users will take up an API call, hence the initial
`Net::HTTP.get_response` that is made first.  This might slow down the
call to this method overall, but it saves API calls in case someone is
providing large commit messages with many code examples included (what
this will be used for in a future commit)