Add CommitMetadataChecker
Checks commits for possible @username additions, and raises errors for
any valid offenses.

Uses GitHub username rules, and will perform the least number of github
username lookups possible to warn for usernames.  Does not outright call
offenses errors, as code examples could be used in the commit messages,
however, ideally if a message is shown it would be courteous to change
the code example to something that doesn't ping a user's email on every
commit update.

Fixes #414