[GitService::Credentials] Add from_ssh_config
Adds GitService::SshConfig, which is used to parse a SSH identity file
from a ~/.ssh/confg (or equivalents) to be used for git operations.

GitService::Credentials.from_ssh_config uses that method to pull out a
config for a given host, and will use the `User` field for that config
if one isn't provided in the URL (which is the `username` arg in this
class).