[GitService::Credentials] Add find_for_user_and_host
This is a generic method for finding a set of credentials to use based
on hostname and uri for a given git url (usually from a remote).

Favors from config, but then falls back to ssh-agent if that doesn't
exist, but for a prod deployment, it probably is better to use the
config.